deployment/scripts/requirements.txt
boto3
requests
ansible
neo4j-driver